Create whatsapp_wa-contacts.py
Be sure to have the wa.db in the script folder.
The script allows users to search for contacts based on phone numbers or specific words within the 'Display Name' or 'WhatsApp Name' columns.
The script is case-insensitive, meaning it ignores the letter case when searching for terms,  and it also normalizes phone numbers by removing any spaces to ensure accurate matching.